+++ src/main/java/com/bbn/hadder/commands/fun/ClydeCommand.java
@@ -8,8 +8,8 @@ import club.minnced.discord.webhook.WebhookClient;
 import club.minnced.discord.webhook.WebhookClientBuilder;
 import com.bbn.hadder.commands.Command;
 import com.bbn.hadder.commands.CommandEvent;
-import com.bbn.hadder.commands.Perm;
-import com.bbn.hadder.commands.Perms;
+import com.bbn.hadder.core.Perm;
+import com.bbn.hadder.core.Perms;
 import com.bbn.hadder.utils.MessageEditor;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Icon;
