+++ src/main/java/com/bbn/hadder/commands/owner/EvalCommand.java
@@ -7,8 +7,8 @@ package com.bbn.hadder.commands.owner;
 import com.bbn.hadder.Hadder;
 import com.bbn.hadder.commands.Command;
 import com.bbn.hadder.commands.CommandEvent;
-import com.bbn.hadder.commands.Perm;
-import com.bbn.hadder.commands.Perms;
+import com.bbn.hadder.core.Perm;
+import com.bbn.hadder.core.Perms;
 import com.bbn.hadder.utils.MessageEditor;
 
 import javax.script.ScriptEngine;
