+++ src/main/java/com/bbn/hadder/commands/owner/RebootCommand.java
@@ -6,8 +6,8 @@ package com.bbn.hadder.commands.owner;
 
 import com.bbn.hadder.commands.Command;
 import com.bbn.hadder.commands.CommandEvent;
-import com.bbn.hadder.commands.Perm;
-import com.bbn.hadder.commands.Perms;
+import com.bbn.hadder.core.Perm;
+import com.bbn.hadder.core.Perms;
 
 @Perms(Perm.BOT_OWNER)
 public class RebootCommand implements Command {
