+++ src/main/java/com/bbn/hadder/core/Perm.java
@@ -2,9 +2,14 @@
  * @author Hax / Hax6775 / Schlauer_Hax
  */
 
-package com.bbn.hadder.commands;
+package com.bbn.hadder.core;
 
+import com.bbn.hadder.commands.CommandEvent;
 import net.dv8tion.jda.api.Permission;
+
+/*
+ * @author Skidder / GregTCLTK
+ */
 
 public enum Perm {
