+++ src/main/java/com/bbn/hadder/core/Perms.java
@@ -2,10 +2,14 @@
  * @author Hax / Hax6775 / Schlauer_Hax
  */
 
-package com.bbn.hadder.commands;
+package com.bbn.hadder.core;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
+
+/*
+ * @author Skidder / GregTCLTK
+ */
 
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Perms {
